+++ package.json
@@ -21,8 +21,7 @@
     "expo-status-bar": "~1.6.0",
     "react": "18.2.0",
     "react-native": "0.72.6",
-    "react-native-svg": "13.9.0",
-    "react-native-svg-uri": "^1.2.3"
+    "react-native-svg": "13.9.0"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0"
